split out remote index creation helpers & renamed
no code changes
