+++ lib/hdk_graph_helpers/src/entry_helpers.rs
@@ -3,7 +3,7 @@
  *
  * @see     ../README.md
  * @package HDK Graph Helpers
- * @since   2020-01-03
+ * @since   2019-05-16
  */
 use std::convert::{ TryFrom };
 use hdk::{
