+++ lib/hdk_graph_helpers/src/key_helpers.rs
@@ -12,7 +12,7 @@
  *
  * @see     ../README.md
  * @package HDK Graph Helpers
- * @since   2020-01-03
+ * @since   2019-05-16
  */
 use hdk::{
     holochain_persistence_api::cas::content::Address,
