+++ lib/hdk_graph_helpers/src/lib.rs
@@ -14,6 +14,7 @@ pub mod record_interface;
 pub mod entry_helpers;
 pub mod key_helpers;
 pub mod local_index_helpers;
+pub mod remote_index_helpers;
 pub mod record_helpers;
 pub mod link_helpers;
 pub mod rpc_helpers;
@@ -25,6 +26,7 @@ pub use maybe_undefined::MaybeUndefined as MaybeUndefined;
 use entry_helpers as entries;
 pub use key_helpers as keys;
 pub use local_index_helpers as local_indexes;
+pub use remote_index_helpers as remote_indexes;
 pub use record_helpers as records;
 pub use link_helpers as links;
 pub use rpc_helpers as rpc;
